spike: cr feedback
